requirements.txt
aiohttp
python-dotenv
requests
beautifulsoup4
loguru
telethon
flask
